backend/requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.2
djangorestframework==3.15.2
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mypy-extensions==1.0.0
mysqlclient==2.2.5
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
